[BuildSystem] Make ShellCommand internal.
 - This is necessary for extensions, which are going to want access to the
   internals of the shell comamnd (and I don't think it is worth making a
   protocol for this).
