+++ lib/BuildSystem/BuildSystem.cpp
@@ -27,6 +27,7 @@
 #include "llbuild/BuildSystem/BuildNode.h"
 #include "llbuild/BuildSystem/BuildValue.h"
 #include "llbuild/BuildSystem/ExternalCommand.h"
+#include "llbuild/BuildSystem/ShellCommand.h"
 #include "llbuild/Core/BuildDB.h"
 #include "llbuild/Core/BuildEngine.h"
 #include "llbuild/Core/DependencyInfoParser.h"
@@ -1924,401 +1925,6 @@ public:
 };
 
 #pragma mark - ShellTool implementation
-
-class ShellCommand : public ExternalCommand {
-  /// The dependencies style to expect (in the `depsPath`).
-  enum class DepsStyle {
-    /// No discovered dependencies are in use.
-    Unused = 0,
-      
-    /// "Makefile" style dependencies in the form typically generated by C
-    /// compilers, wherein the dependencies of the first target are treated as
-    /// dependencies of the command.
-    Makefile,
-
-    /// Darwin's DependencyInfo format.
-    DependencyInfo,
-  };
-  
-  /// The command line arguments.
-  std::vector<StringRef> args;
-
-  /// Arbitrary string used to contribute to the task signature.
-  std::string signatureData;
-
-  /// The environment to use. If empty, the environment will be inherited.
-  SmallVector<std::pair<StringRef, StringRef>, 1> env;
-  
-  /// The path to the dependency output file, if used.
-  SmallVector<std::string, 1> depsPaths{};
-
-  /// The style of dependencies used.
-  DepsStyle depsStyle = DepsStyle::Unused;
-
-  /// Whether to inherit the base environment.
-  bool inheritEnv = true;
-
-  /// Whether it is safe to interrupt (SIGINT) the tool during cancellation.
-  bool canSafelyInterrupt = true;
-
-  /// Working directory in which to spawn the external command
-  std::string workingDirectory;
-
-  /// Whether the control pipe is enabled for this command
-  bool controlEnabled = true;
-
-  /// The cached signature, once computed -- 0 is used as a sentinel value.
-  std::atomic<CommandSignature> cachedSignature{ };
-  
-  virtual CommandSignature getSignature() override {
-    CommandSignature signature = cachedSignature;
-    if (!signature.isNull())
-      return signature;
-
-    auto code = ExternalCommand::getSignature();
-    if (!signatureData.empty()) {
-      code = code.combine(signatureData);
-    } else {
-      for (const auto& arg: args) {
-        code = code.combine(arg);
-      }
-      for (const auto& entry: env) {
-        code = code.combine(entry.first);
-        code = code.combine(entry.second);
-      }
-      for (const auto& path: depsPaths) {
-        code = code.combine(path);
-      }
-      code = code.combine(int(depsStyle));
-      code = code.combine(int(inheritEnv));
-      code = code.combine(int(canSafelyInterrupt));
-    }
-    signature = code;
-    if (signature.isNull()) {
-      signature = CommandSignature(1);
-    }
-    cachedSignature = signature;
-    return signature;
-  }
-
-  bool processDiscoveredDependencies(BuildSystemCommandInterface& bsci,
-                                     Task* task,
-                                     QueueJobContext* context) {
-    // It is an error if the dependencies style is not specified.
-    //
-    // FIXME: Diagnose this sooner.
-    if (depsStyle == DepsStyle::Unused) {
-      getBuildSystem(bsci.getBuildEngine()).error(
-          "", "missing required 'deps-style' specifier");
-      return false;
-    }
-
-    for (const auto& depsPath: depsPaths) {
-      // Read the dependencies file.
-      auto input = bsci.getFileSystem().getFileContents(depsPath);
-      if (!input) {
-        getBuildSystem(bsci.getBuildEngine()).getDelegate().commandHadError(this,
-                       "unable to open dependencies file (" + depsPath + ")");
-        return false;
-      }
-
-      switch (depsStyle) {
-      case DepsStyle::Unused:
-        assert(0 && "unreachable");
-        break;
-
-      case DepsStyle::Makefile:
-        if (!processMakefileDiscoveredDependencies(
-                bsci, task, context, depsPath, input.get()))
-          return false;
-        continue;
-
-      case DepsStyle::DependencyInfo:
-        if (!processDependencyInfoDiscoveredDependencies(
-                bsci, task, context, depsPath, input.get()))
-          return false;
-        continue;
-      }
-      
-      llvm::report_fatal_error("unknown dependencies style");
-    }
-
-    return true;
-  }
-
-  bool processMakefileDiscoveredDependencies(BuildSystemCommandInterface& bsci,
-                                             Task* task,
-                                             QueueJobContext* context,
-                                             StringRef depsPath,
-                                             llvm::MemoryBuffer* input) {
-    // Parse the output.
-    //
-    // We just ignore the rule, and add any dependency that we encounter in the
-    // file.
-    struct DepsActions : public core::MakefileDepsParser::ParseActions {
-      BuildSystemCommandInterface& bsci;
-      Task* task;
-      ShellCommand* command;
-      StringRef depsPath;
-      unsigned numErrors{0};
-
-      DepsActions(BuildSystemCommandInterface& bsci, Task* task,
-                  ShellCommand* command, StringRef depsPath)
-          : bsci(bsci), task(task), command(command), depsPath(depsPath) {}
-
-      virtual void error(const char* message, uint64_t position) override {
-        std::stringstream msg;
-        msg << "error reading dependency file '" << depsPath.str() << "': "
-            << message << " at position " << position;
-        getBuildSystem(bsci.getBuildEngine()).getDelegate().commandHadError(
-            command, msg.str());
-        ++numErrors;
-      }
-
-      virtual void actOnRuleDependency(const char* dependency,
-                                       uint64_t length,
-                                       const StringRef unescapedWord) override {
-        if (llvm::sys::path::is_absolute(unescapedWord)) {
-          bsci.taskDiscoveredDependency(task, BuildKey::makeNode(unescapedWord));
-          return;
-        }
-
-        // Generate absolute path
-        //
-        // NOTE: This is making the assumption that relative paths coming in a
-        // dependency file are in relation to the explictly set working
-        // directory, or the current working directory when it has not been set.
-        SmallString<PATH_MAX> absPath = StringRef(command->workingDirectory);
-        llvm::sys::path::append(absPath, unescapedWord);
-        llvm::sys::fs::make_absolute(absPath);
-
-        bsci.taskDiscoveredDependency(task, BuildKey::makeNode(absPath));
-      }
-
-      virtual void actOnRuleStart(const char* name, uint64_t length,
-                                  const StringRef unescapedWord) override {}
-
-      virtual void actOnRuleEnd() override {}
-    };
-
-    DepsActions actions(bsci, task, this, depsPath);
-    core::MakefileDepsParser(input->getBufferStart(), input->getBufferSize(),
-                             actions).parse();
-    return actions.numErrors == 0;
-  }
-
-  bool
-  processDependencyInfoDiscoveredDependencies(BuildSystemCommandInterface& bsci,
-                                              Task* task,
-                                              QueueJobContext* context,
-                                              StringRef depsPath,
-                                              llvm::MemoryBuffer* input) {
-    // Parse the output.
-    //
-    // We just ignore the rule, and add any dependency that we encounter in the
-    // file.
-    struct DepsActions : public core::DependencyInfoParser::ParseActions {
-      BuildSystemCommandInterface& bsci;
-      Task* task;
-      ShellCommand* command;
-      StringRef depsPath;
-      unsigned numErrors{0};
-
-      DepsActions(BuildSystemCommandInterface& bsci, Task* task,
-                  ShellCommand* command, StringRef depsPath)
-          : bsci(bsci), task(task), command(command), depsPath(depsPath) {}
-
-      virtual void error(const char* message, uint64_t position) override {
-        getBuildSystem(bsci.getBuildEngine()).getDelegate().commandHadError(command,
-                       "error reading dependency file '" + depsPath.str() +
-                       "': " + std::string(message));
-        ++numErrors;
-      }
-
-      // Ignore everything but actual inputs.
-      virtual void actOnVersion(StringRef) override { }
-      virtual void actOnMissing(StringRef) override { }
-      virtual void actOnOutput(StringRef) override { }
-
-      virtual void actOnInput(StringRef name) override {
-        bsci.taskDiscoveredDependency(task, BuildKey::makeNode(name));
-      }
-    };
-
-    DepsActions actions(bsci, task, this, depsPath);
-    core::DependencyInfoParser(input->getBuffer(), actions).parse();
-    return actions.numErrors == 0;
-  }
-
-public:
-  using ExternalCommand::ExternalCommand;
-  ShellCommand(StringRef name, bool controlEnabled) : ExternalCommand(name),
-    controlEnabled(controlEnabled) { }
-
-  virtual void getShortDescription(SmallVectorImpl<char> &result) const override {
-    llvm::raw_svector_ostream(result) << getDescription();
-  }
-
-  virtual void getVerboseDescription(SmallVectorImpl<char> &result) const override {
-    llvm::raw_svector_ostream os(result);
-    bool first = true;
-    for (const auto& arg: args) {
-      if (!first) os << " ";
-      first = false;
-      basic::appendShellEscapedString(os, arg);
-    }
-  }
-  
-  virtual bool configureAttribute(const ConfigureContext& ctx, StringRef name,
-                                  StringRef value) override {
-    if (name == "args") {
-      // When provided as a scalar string, we default to executing using the
-      // shell.
-      args.clear();
-#if defined(_WIN32)
-      args.push_back(ctx.getDelegate().getInternedString(
-          "C:\\windows\\system32\\cmd.exe"));
-      args.push_back(ctx.getDelegate().getInternedString("/C"));
-#else
-      args.push_back(ctx.getDelegate().getInternedString("/bin/sh"));
-      args.push_back(ctx.getDelegate().getInternedString("-c"));
-#endif
-      args.push_back(ctx.getDelegate().getInternedString(value));
-    } else if (name == "signature") {
-      signatureData = value;
-    } else if (name == "deps") {
-      depsPaths.clear();
-      depsPaths.emplace_back(value);
-    } else if (name == "deps-style") {
-      if (value == "makefile") {
-        depsStyle = DepsStyle::Makefile;
-      } else if (value == "dependency-info") {
-        depsStyle = DepsStyle::DependencyInfo;
-      } else {
-        ctx.error("unknown 'deps-style': '" + value + "'");
-        return false;
-      }
-      return true;
-    } else if (name == "can-safely-interrupt") {
-      if (value != "true" && value != "false") {
-        ctx.error("invalid value: '" + value + "' for attribute '" +
-                  name + "'");
-        return false;
-      }
-      canSafelyInterrupt = value == "true";
-    } else if (name == "inherit-env") {
-      if (value != "true" && value != "false") {
-        ctx.error("invalid value: '" + value + "' for attribute '" +
-                  name + "'");
-        return false;
-      }
-      inheritEnv = value == "true";
-    } else if (name == "working-directory") {
-      // Ensure the working directory is absolute. This will make sure any
-      // relative directories are interpreted as relative to the CWD at the time
-      // the rule is defined.
-      SmallString<PATH_MAX> wd = value;
-      llvm::sys::fs::make_absolute(wd);
-      workingDirectory = StringRef(wd);
-    } else if (name == "control-enabled") {
-      if (value != "true" && value != "false") {
-        ctx.error("invalid value: '" + value + "' for attribute '" +
-                  name + "'");
-        return false;
-      }
-      controlEnabled = value == "true";
-    } else {
-      return ExternalCommand::configureAttribute(ctx, name, value);
-    }
-
-    return true;
-  }
-  
-  virtual bool configureAttribute(const ConfigureContext& ctx, StringRef name,
-                                  ArrayRef<StringRef> values) override {
-    if (name == "args") {
-      // Diagnose missing arguments.
-      if (values.empty()) {
-        ctx.error("invalid arguments for command '" + getName() + "'");
-        return false;
-      }
-
-      args.clear();
-      args.reserve(values.size());
-      for (auto arg: values) {
-        args.emplace_back(ctx.getDelegate().getInternedString(arg));
-      }
-    } else if (name == "deps") {
-      depsPaths.clear();
-      depsPaths.insert(depsPaths.begin(), values.begin(), values.end());
-    } else {
-      return ExternalCommand::configureAttribute(ctx, name, values);
-    }
-
-    return true;
-  }
-
-  virtual bool configureAttribute(
-      const ConfigureContext& ctx, StringRef name,
-      ArrayRef<std::pair<StringRef, StringRef>> values) override {
-    if (name == "env") {
-      env.clear();
-      env.reserve(values.size());
-      for (const auto& entry: values) {
-        env.emplace_back(
-            std::make_pair(
-                ctx.getDelegate().getInternedString(entry.first),
-                ctx.getDelegate().getInternedString(entry.second)));
-      }
-    } else {
-      return ExternalCommand::configureAttribute(ctx, name, values);
-    }
-
-    return true;
-  }
-
-  virtual void executeExternalCommand(
-      BuildSystemCommandInterface& bsci,
-      Task* task,
-      QueueJobContext* context,
-      llvm::Optional<ProcessCompletionFn> completionFn) override {
-
-    // Execute the command.
-    bsci.getExecutionQueue().executeProcess(
-        context, args, env,
-        /*inheritEnvironment=*/inheritEnv,
-        {canSafelyInterrupt, workingDirectory, controlEnabled},
-        /*completionFn=*/{[this, &bsci, task, completionFn](ProcessResult result) {
-          if (result.status != ProcessStatus::Succeeded) {
-            // If the command failed, there is no need to gather dependencies.
-            if (completionFn.hasValue())
-              completionFn.getValue()(result);
-            return;
-          }
-
-          // Collect the discovered dependencies, if used.
-          if (!depsPaths.empty()) {
-            // FIXME: Really want this job to go into a high priority fifo queue
-            // so as to not hold up downstream tasks.
-            bsci.addJob({ this, [this, &bsci, task, completionFn, result](QueueJobContext* context) {
-              if (!processDiscoveredDependencies(bsci, task, context)) {
-                // If we were unable to process the dependencies output, report a
-                // failure.
-                if (completionFn.hasValue())
-                  completionFn.getValue()(ProcessStatus::Failed);
-                return;
-              }
-              if (completionFn.hasValue())
-                completionFn.getValue()(result);
-            }});
-            return;
-          }
-
-          if (completionFn.hasValue())
-            completionFn.getValue()(result);
-        }});
-  }
-};
 
 class ShellTool : public Tool {
 private:
